pyproject.toml
[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[project]
name = "code-context-agent"
version = "3.1.0"
description = "A CLI tool for code context analysis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-protocol>=0.1.13",
    "ag-ui-strands>=0.1.1",
    "cyclopts>=4.6.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "networkx[default,extra]>=3.6.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "rich>=14.2.0",
    "strands-agents>=1.22.0",
    "strands-agents-tools>=0.2.19",
    "ty>=0.0.18",
]

[project.scripts]
code-context-agent = "code_context_agent.cli:app"

[dependency-groups]
dev = [
    "commitizen>=4.13.8",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.15.2",
    "strands-agents-evals>=0.1.7",
]
security = [
    "bandit>=1.9.3",
    "pip-audit>=2.10.0",
]

# =============================================================================
# Ruff Configuration (Linting & Formatting)
# =============================================================================

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py313"
exclude = [".git", ".venv", "__pycache__", "build", "dist", ".ruff_cache"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "C90",    # mccabe complexity
    "BLE",    # flake8-blind-except (catch specific exceptions)
    "COM",    # flake8-commas (enforce trailing commas)
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # pylint
    "RUF",    # Ruff-specific
    "D",      # pydocstyle (docstrings)
    "S",      # flake8-bandit (security)
]
ignore = [
    "D100",    # Allow missing module docstrings
    "D104",    # Allow missing package docstrings
    "D301",    # Allow backslashes in docstrings (for LSP protocol docs)
    "PLC0415", # Allow imports inside functions (for circular import avoidance)
    "B027",    # Allow empty methods in ABCs (optional hook methods)
    "S602",    # Allow subprocess shell=True (needed for shell commands)
    "S603",    # Allow subprocess calls (needed for shell tool execution)
    "S607",    # Allow partial executable paths (sh -c pattern is safe)
    "PLR0913", # Allow many function arguments (tool function design)
    "PLW1510", # Allow subprocess.run without check (we handle errors)
    "ARG002",  # Allow unused method arguments (protocol interface compliance)
    "SIM105",  # Allow try-except-pass patterns (explicit is clearer)
    "G004",    # Allow f-strings in logging (using loguru)
    "TID252",  # Allow relative imports (project style choice)
    "EM101",   # Allow string literals in exceptions
    "EM102",   # Allow f-strings in exceptions
    "TRY003",  # Allow long exception messages
]
fixable = ["ALL"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10  # Enforce complexity limit after refactoring

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "S101", "ARG"]
"**/__init__.py" = ["F401", "RUF022"]  # Allow unused imports and unsorted __all__ in init files
"src/code_context_agent/tools/**/*.py" = ["PLR2004", "PLR0911", "PLR0912", "SIM102", "SIM108"]  # Allow magic numbers and dispatcher patterns in tools

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true

# =============================================================================
# Ty Configuration (Type Checking)
# =============================================================================

[tool.ty.rules]
possibly-unresolved-reference = "error"
invalid-argument-type = "error"
# incompatible-assignment - not yet supported by ty
missing-argument = "error"
unsupported-operator = "error"
division-by-zero = "error"
unused-ignore-comment = "warn"
redundant-cast = "warn"

[tool.ty.environment]
python-version = "3.13"

[tool.ty.src]
include = ["src"]

# =============================================================================
# Pytest Configuration
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# =============================================================================
# Commitizen Configuration (Conventional Commits)
# =============================================================================

[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "pep621"
version_files = [
    "src/code_context_agent/__init__.py:__version__"
]
tag_format = "v$version"
update_changelog_on_bump = true
version_scheme = "semver"
